Add files via upload
Dragon on castle- edited illustration- converted it into a line drawing in white ink and added in all of the relevant details;  I also added a castle, a new background and an egg as the last step. All of this was done on an iPad mini via Notepad and a sketch app.